Added more links and changes
