+++ Maintenance/Education page hindi/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="wdth=devce-width , initial-scale=1.0">
+    <title>Maintainence Instructional Video</title>
+
+    <link rel="stylesheet" href="style.css">
+
+</head>
+<body>
+
+    <div class="container">
+        <div class="main-video-container">
+            <video src="media/vid1.mp4" loop controls class="main-video"></video>
+            <h3 class="main-vid-title">ट्रक रखरखाव शिक्षात्मक वीडियो</h3>
+        </div>
+        <div class="video-list-container" id="videosList"></div>
+    </div>
+
+    <script src="Videoplaylist.js"></script>
+    
+
+</body>
+</html>
