--- flask_AI/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="styles.css">
-    <title>Truck Maintenance Prediction</title>
-</head>
-<body>
-    <div class="container">
-        <h1>Truck Maintenance Prediction</h1>
-        <input type="number" id="operation-time" placeholder="Operation Time">
-        <input type="text" id="location" placeholder="Location (desert/hilly/rainy/humid/all)">
-        <input type="text" id="company" placeholder="Company (tata/mahindra/ashok leyland/bharat benz/eicher motors)">
-        <input type="text" id="engine" placeholder="Engine">
-        <input type="number" id="milage" placeholder="Milage">
-        <button id="predict-btn">Get Prediction</button>
-        <div id="prediction-result"></div>
-    </div>
-    <script src="script.js"></script>
-</body>
-</html>
